fixed: some skill-bonuses

+++ _workingstuff/extract_data.rb
@@ -986,7 +986,7 @@ class MagicObjectEntry < Table
     def initialize(csv_row)
         super(csv_row)
         @bonus = BonusStuff.new(csv_row)
-        @skilllvarg = csv_row['ability_skilllvarg'].to_i
+        @skilllvarg = csv_row['ability_skilllvarg'].to_f
     end
 
     def SkipThisItem?
